# Adds HostFunctionBuilder to enable high performance host functions