--- a/internal/wasm/binary/code.go
+++ b/internal/wasm/binary/code.go
@@ -84,7 +84,7 @@ func decodeCode(r *bytes.Reader) (*wasm.Code, error) {
 // See https://www.w3.org/TR/2019/REC-wasm-core-1-20191205/#binary-code
 func encodeCode(c *wasm.Code) []byte {
 	if c.GoFunc != nil {
-		panic("BUG: GoFunc is not encodable")
+		panic("BUG: GoFunction is not encodable")
 	}
 
 	// local blocks compress locals while preserving index order by grouping locals of the same type.