… builder_test.go
… examples/allocation/rust/greet.go
… examples/allocation/tinygo/greet.go
… examples/allocation/zig/greet.go
… examples/import-go/age-calculator.go
… examples/multiple-results/multiple-results.go
… examples/namespace/counter.go
… experimental/listener_example_test.go
… experimental/logging/log_listener_test.go
… imports/assemblyscript/assemblyscript_example_test.go
… imports/emscripten/emscripten_example_test.go
… imports/go/gojs.go
… imports/wasi_snapshot_preview1/args.go
… imports/wasi_snapshot_preview1/args_test.go
… imports/wasi_snapshot_preview1/clock.go
… imports/wasi_snapshot_preview1/clock_test.go
… imports/wasi_snapshot_preview1/environ.go
… imports/wasi_snapshot_preview1/environ_test.go
… imports/wasi_snapshot_preview1/fs.go
… imports/wasi_snapshot_preview1/fs_test.go
… imports/wasi_snapshot_preview1/poll.go
… imports/wasi_snapshot_preview1/random.go
… imports/wasi_snapshot_preview1/testdata/wasi_arg.wat
… imports/wasi_snapshot_preview1/wasi.go
… internal/engine/compiler/compiler_test.go
… internal/engine/compiler/engine.go
… internal/engine/compiler/engine_test.go
… internal/engine/interpreter/interpreter.go
… internal/gojs/crypto_test.go
… internal/gojs/runtime.go