@codefromthecrypt 2022-10-28
the integrated performance results are phenominally better for fine-grained ABI:

ex. https://github.com/http-wasm/http-wasm-host-go/pull/38 run via https://github.com/http-wasm/http-wasm-guest-tinygo benchmarks, TinyGo req/response logging went down from 43.5us/op to 14.5us/op

Before
```
Benchmark/example_router_host_response/TinyGo-16     	  352765	      3334 ns/op
Benchmark/example_router_host_response/wat-16        	  363664	      3171 ns/op
Benchmark/example_router_wasm_response/wat-16        	  232335	      4712 ns/op
Benchmark/example_router_wasm_response/TinyGo-16     	  225828	      4966 ns/op
Benchmark/example_wasi/TinyGo-16                     	   27218	     43631 ns/op
Benchmark/example_wasi/wat-16                        	   36333	     31908 ns/op
```

After
```
Benchmark/example_router_host_response/TinyGo-16  	 1148056	       979.8 ns/op
Benchmark/example_router_host_response/wat-16     	 1269897	       886.1 ns/op
Benchmark/example_router_wasm_response/wat-16     	 1098566	      1014 ns/op
Benchmark/example_router_wasm_response/TinyGo-16  	  802510	      1416 ns/op
Benchmark/example_wasi/TinyGo-16                  	   77762	     14477 ns/op
Benchmark/example_wasi/wat-16                     	  176056	      5794 ns/op
```