@codefromthecrypt 2022-10-28
removed enforcement of context param when using reflection later per #832